즐겨찾기 추가 시나리오 및 기능 개발
