--- src/main/java/nextstep/favorite/controller/dto/FavoriteResponse.java
@@ -1,7 +1,0 @@
-package nextstep.favorite.controller.dto;
-
-/**
- * TODO: StationResponse를 포함하는 클래스로 만듭니다.
- */
-public class FavoriteResponse {
-}
