--- src/main/java/nextstep/favorite/domain/repository/FavoriteRepository.java
@@ -1,7 +1,0 @@
-package nextstep.favorite.domain.repository;
-
-import nextstep.favorite.domain.entity.Favorite;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
-}
