+++ src/main/java/nextstep/member/infrastructure/jwt/JwtTokenProvider.java
@@ -28,7 +28,11 @@ public class JwtTokenProvider implements TokenGenerator {
     }
 
     public String getPrincipal(String token) {
-        return Jwts.parser().setSigningKey(secretKey).parseClaimsJws(token).getBody().getSubject();
+        return Jwts.parser()
+                .setSigningKey(secretKey)
+                .parseClaimsJws(token)
+                .getBody()
+                .getSubject();
     }
 
     public boolean validateToken(String token) {
