+++ src/main/java/nextstep/subway/domain/command/FavoriteCommand.java
@@ -1,0 +1,17 @@
+package nextstep.subway.domain.command;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.ToString;
+
+public class FavoriteCommand {
+
+    @ToString
+    @Getter
+    @AllArgsConstructor
+    public static class CreateFavorite {
+        private Long memberId;
+        private Long source;
+        private Long target;
+    }
+}
