+++ src/main/java/nextstep/subway/domain/exception/NotFoundFavoriteException.java
@@ -1,0 +1,7 @@
+package nextstep.subway.domain.exception;
+
+public class NotFoundFavoriteException extends SubwayDomainException {
+    public NotFoundFavoriteException(Long id) {
+        super(SubwayDomainExceptionType.NOT_FOUND_LINE, "favoriteId " + id + " is not found");
+    }
+}
