--- src/test/java/nextstep/favorite/acceptance/FavoriteAcceptanceTest.java
@@ -1,26 +1,0 @@
-package nextstep.favorite.acceptance;
-
-import nextstep.subway.setup.BaseTestSetup;
-import org.junit.jupiter.api.DisplayName;
-
-@DisplayName("즐겨찾기 관련 기능")
-public class FavoriteAcceptanceTest extends BaseTestSetup {
-
-    /**
-     * Given: 지하철 노선이 등록되어 있고
-     * When: 회원이 경로를 즐겨찾기로 등록하면
-     * Then: 해당 경로가 즐겨찾기로 등록되고 즐겨찾기 목록에 포함된다.
-     */
-
-    /**
-     * Given: 회원이 2개의 즐겨찾기를 등록했고
-     * When: 본인의 즐켜찾기 목록을 조회하면
-     * Then: 2개의 즐겨찾기 목록이 반환된다.
-     */
-
-    /**
-     * Given: 회원이 즐겨찾기를 등록했고
-     * When: 본인의 즐켜찾기 목록을 삭제하면
-     * Then: 해당 즐켜찾기가 삭제되고 즐겨찾기 목록에서 제외된다.
-     */
-}
