+++ src/test/java/nextstep/member/acceptance/AuthAcceptanceTest.java
@@ -5,16 +5,17 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.member.domain.entity.Member;
 import nextstep.member.domain.repository.MemberRepository;
-import nextstep.subway.setup.BaseTestSetup;
+import nextstep.util.BaseTestSetup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 import java.util.HashMap;
 import java.util.Map;
 
+import static nextstep.member.acceptance.AuthSteps.로그인;
+import static nextstep.member.acceptance.AuthSteps.인증토큰을_추출한다;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class AuthAcceptanceTest extends BaseTestSetup {
@@ -30,18 +31,7 @@ class AuthAcceptanceTest extends BaseTestSetup {
     void bearerAuth() {
         memberRepository.save(new Member(EMAIL, PASSWORD, AGE));
 
-        Map<String, String> params = new HashMap<>();
-        params.put("email", EMAIL);
-        params.put("password", PASSWORD);
-
-        ExtractableResponse<Response> response = RestAssured.given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(params)
-                .when().post("/login/token")
-                .then().log().all()
-                .statusCode(HttpStatus.OK.value()).extract();
-
-        String accessToken = response.jsonPath().getString("accessToken");
+        String accessToken = 인증토큰을_추출한다(로그인(EMAIL, PASSWORD));
         assertThat(accessToken).isNotBlank();
 
         ExtractableResponse<Response> response2 = RestAssured.given().log().all()
