+++ src/test/java/nextstep/subway/acceptance/SectionAcceptanceTest.java
@@ -1,6 +1,6 @@
 package nextstep.subway.acceptance;
 
-import nextstep.subway.setup.BaseTestSetup;
+import nextstep.util.BaseTestSetup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
