+++ src/test/java/nextstep/subway/fixtures/LineFixture.java
@@ -2,6 +2,7 @@ package nextstep.subway.fixtures;
 
 import nextstep.subway.domain.entity.line.Line;
 import nextstep.subway.domain.entity.line.LineSections;
+import nextstep.subway.domain.entity.station.Station;
 
 import java.util.UUID;
 import java.util.stream.IntStream;
@@ -18,6 +19,14 @@ public class LineFixture {
         IntStream
                 .range(0, upDownStationIds.length-1)
                 .forEach((i) -> line.addSection(upDownStationIds[i], upDownStationIds[i+1], 10L));
+        return line;
+    }
+
+    public static Line prepareConnectedLine(Station... upDownStations) {
+        Line line = new Line(UUID.randomUUID().toString(), UUID.randomUUID().toString(), new LineSections());
+        IntStream
+                .range(0, upDownStations.length-1)
+                .forEach((i) -> line.addSection(upDownStations[i].getId(), upDownStations[i+1].getId(), 10L));
         return line;
     }
 }
